+++ packages/semi-ui-vue/src/components/radio/radio.tsx
@@ -36,7 +36,7 @@ export type RadioProps = {
   autoFocus?: boolean;
   checked?: boolean;
   defaultChecked?: boolean;
-  value?: string | number;
+  value?: string | number | boolean;
   disabled?: boolean;
   prefixCls?: string;
   displayMode?: RadioDisplayMode;
@@ -79,7 +79,7 @@ export const vuePropsType: CombineProps<RadioProps> = {
     type: Boolean,
     default: false,
   },
-  value: [String, Number],
+  value: [String, Number, Boolean],
   disabled: Boolean,
   prefixCls: String,
   displayMode: String as PropType<RadioProps['displayMode']>,
@@ -289,13 +289,11 @@ const Radio = defineComponent({
           {...getDataAttr()}
         >
           <RadioInner
-            {...props}
-            {...props_}
+            {...({...props, ...props_, onChange: onChange})}
             mode={realMode}
             name={name ?? groupName}
             isButtonRadio={isButtonRadio}
             isPureCardRadioGroup={isPureCardRadioGroup}
-            onChange={onChange}
             ref={(ref: any) => {
               radioEntity = ref;
             }}